fix(SIWA): fix review notes
